+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added examples how to use the low-level and high-level APIs.
 - Added backwards compability in floor reader and writer for lists as produced by Athena.
 - Fixed many crash issues found during fuzzing of the parquet reader.